move tetrion_snapshot to recording folder
